migrate: courses page
