+++ src/components/PageComp/CoursesPage/About.jsx
