+++ src/components/PageComp/CoursesPage/index.js
@@ -1,0 +1,3 @@
+export { default as About } from './About';
+export { default as Courses } from './Courses';
+export { default as Hero } from './Hero';
