--- src/components/coursespage/Hero/index.module.css
@@ -1,14 +1,0 @@
-.section {
-  text-align: center;
-  background-image: rgb(var(--base-100), 0.9);
-}
-.layout {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.layout p {
-  margin-top: 1rem;
-}
