--- src/components/coursespage/about/index.module.css
@@ -1,42 +1,0 @@
-.layout {
-  padding-block: 6rem;
-  margin-inline: auto;
-}
-.cards {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
-.card {
-  padding: 1rem /* 16px */;
-  padding-block: 2rem;
-  background-image: linear-gradient(
-    to right,
-    rgb(var(--content), 0),
-    rgb(var(--content), 0.1)
-  );
-  border-radius: 0.5rem;
-  border: 1px solid rgb(var(--content), 0.2);
-  text-align: center;
-}
-
-@media (min-width: 1024px) {
-  .card {
-    margin-bottom: 0px;
-    flex: 1;
-    padding: 2rem;
-  }
-}
-@media (min-width: 871px) {
-  .card {
-    width: 49%;
-  }
-}
-.card p {
-  margin-block: 1rem;
-}
-.section .btn {
-  border-radius: 10rem;
-  padding: 0.6rem 1rem;
-  margin-top: 1rem;
-}
