--- src/components/coursespage/index.js
@@ -1,4 +1,0 @@
-export { default as About } from './about';
-export { default as Courses } from './courses';
-export { default as Hero } from './Hero';
-export { default as Testimonials } from './Testimonial';
