+++ src/pages/CoursesPage.jsx
@@ -1,5 +1,5 @@
-import { About, Courses, Hero, Testimonials } from '@/components/coursespage';
-import Layout from '@/components/layout';
+import Layout, { Testimonials } from '@/components/layout';
+import { About, Courses, Hero } from '@/components/PageComp/CoursesPage';
 
 import { CoursesPageContent } from './content';
 
@@ -9,7 +9,17 @@ const CoursesPage = () => {
       <Hero />
       <About />
       <Courses />
-      <Testimonials />
+      <Testimonials
+        tweetId={[
+          '1620824479395090435',
+          '1620081639199244288',
+          '1620489706994806785',
+          '1620828035804184579',
+          '1617535692497883138',
+          '1487399069383155716',
+          '1544248976664567811',
+        ]}
+      />
     </Layout>
   );
 };
